# apps/api/openapi.yaml
openapi: "3.1.0"
info:
  title: ATHENA M&E Platform API
  version: "0.1.0"
  description: |
    Backend API for the ATHENA Monitoring & Evaluation platform.
    All endpoints (except /health and /auth/*) require a valid
    JWT issued via OAuth SSO, carried as an HttpOnly cookie
    (`athena_session`) or an `Authorization: Bearer <token>` header.
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://your.domain.com
    description: Production

# ── Security scheme ───────────────────────────────────────────
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: athena_session

  # ── Reusable schemas ────────────────────────────────────────
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UUID:
      type: string
      format: uuid
    ISOTimestamp:
      type: string
      format: date-time
    Role:
      type: string
      enum: [admin, supervisor, enumerator]
...
    User:
      type: object
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        org_id: { $ref: "#/components/schemas/UUID" }
        email: { type: string, format: email }
        display_name: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        is_active: { type: boolean }
        created_at: { $ref: "#/components/schemas/ISOTimestamp" }
    AuthUser:
      type: object
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/Role" }
        org_id: { $ref: "#/components/schemas/UUID" }
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Insufficient role
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
security:
  - sessionCookie: []

# ── Paths ─────────────────────────────────────────────────────
paths:
...
  /auth/callback/{provider}:
    get:
      operationId: authCallback
      summary: OAuth callback — exchange code, set session cookie
      security: []
      tags: [Auth]
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string, enum: [google, microsoft] }
        - name: code
          in: query
          required: true
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
      responses:
        "302":
          description: Redirect to dashboard after successful login
          headers:
            Set-Cookie:
              schema:
                {
                  type: string,
                  description: "athena_session=<jwt>; HttpOnly; SameSite=Strict",
                }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /auth/me:
    get:
      operationId: getMe
      summary: Return current authenticated user
      tags: [Auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ── Organisation ─────────────────────────────────────────────
...
  /org/users:
    get:
      operationId: listUsers
      summary: List all users in the organisation
      tags: [Organisation]
      x-roles: [admin]
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

    post:
      operationId: inviteUser
      summary: Invite a new user (creates inactive record, sends email)
      tags: [Organisation]
      x-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email: { type: string, format: email }
                role: { $ref: "#/components/schemas/Role" }
      responses:
        "201":
          description: User invited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /submissions:
    head:
      operationId: submissionCapability
      summary: OpenRosa capability discovery
      tags: [Ingestion]
      x-roles: [admin, supervisor, enumerator]
      responses:
        "204":
          description: Server is OpenRosa capable
          headers:
            X-OpenRosa-Version:
              schema: { type: string, example: "1.0" }

    post:
      operationId: ingestSubmission
      summary: Ingest a form submission (OpenRosa 1.0 multipart)
      tags: [Ingestion]
      x-roles: [admin, supervisor, enumerator]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [xml_submission_file]
              properties:
                xml_submission_file:
                  type: string
                  format: binary
                  description: XLSForm metadata XML
                payload:
                  type: string
                  description: JSON survey answers
      responses:
        "202":
          description: Submission accepted (written to sector table or quarantined)
        "409":
          description: Conflict detected — written to submission_conflicts
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Payload structure invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # ── Quarantine ────────────────────────────────────────────────
...